+++ crates/wdl-engine/tests/tasks.rs
@@ -166,7 +166,11 @@ fn configs() -> Vec<config::Config> {
     vec![
         {
             config::Config {
-                backend: BackendConfig::Local(Default::default()),
+                backends: [(
+                    "default".to_string(),
+                    BackendConfig::Local(Default::default()),
+                )]
+                .into(),
                 ..Default::default()
             }
         },
@@ -176,7 +180,11 @@ fn configs() -> Vec<config::Config> {
         #[cfg(target_os = "linux")]
         {
             config::Config {
-                backend: BackendConfig::Docker(Default::default()),
+                backends: [(
+                    "default".to_string(),
+                    BackendConfig::Docker(Default::default()),
+                )]
+                .into(),
                 ..Default::default()
             }
         },